refactor: move node specific commands into own module
